add: "unlimited" instruction quantity
Now the computer's instruction pointer is 32-bit, allowing "unlimited" programs

+++ interpreter/Core/VirtualMachine.cs
@@ -10,7 +10,7 @@ namespace interpreter.Core
         public byte[] RAM { get; } = new byte[MAX_RAM];
         public byte[] Registers { get; } = new byte[MAX_REGISTERS];
     
-        public byte IP { get; set; } = 0;
+        public int IP { get; set; } = 0;
         public Stack<byte> CallStack { get; } = new Stack<byte>();
         public byte Output { get; set; } = 0;
